refactor: modify the default mode of the bridge to legacy mode to reduce redundant code
