+++ apps/router-demo/router-remote1-2001/src/export-App.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import App from './App';
-import { createBridgeComponent } from '@module-federation/bridge-react/legacy';
+import { createBridgeComponent } from '@module-federation/bridge-react';
 
 const provider = createBridgeComponent({
   rootComponent: App,
